+++ app/src/main/kotlin/dev/vitalcc/stukay/StukayApplication.kt
@@ -2,4 +2,4 @@ package dev.vitalcc.stukay
 
 import android.app.Application
 
-class StukayApplication : Application()
+public class StukayApplication : Application()